# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TTR ONE — ERP Platform API
  version: 0.1.0
  description: |
    REST API for **TTR ONE**, a multi-tenant SaaS ERP for Uzbekistan
    (Node/Fastify + PostgreSQL/Prisma backend, Nuxt 3 SPA frontend).

    ## Conventions
    - **Base URL**: all endpoints live under `/api/v1`.
    - **Auth**: JWT Bearer. Obtain an `accessToken` from `POST /auth/login`
      (15-minute TTL) and rotate it via `POST /auth/refresh`.
    - **Money**: all monetary amounts are integers in **minor units (tiyin)**,
      e.g. `1 000 000` = 10 000.00 UZS. Fields end in `Minor`.
    - **Quantities**: decimal strings, bounded to 6 decimal places.
    - **Multi-tenancy**: every record is scoped to the caller's tenant
      (derived from the token); there is no tenant id in the path.
    - **Errors**: uniform envelope `{ "error": { "code": "...", "message": "..." } }`.
    - **Pagination**: list endpoints accept `page`, `pageSize` (max 200),
      `sort`, `order` (`asc`/`desc`), `search`, and return a `meta` object
      `{ page, pageSize, total, totalPages }`.
  contact:
    name: TTR ONE
servers:
  - url: http://localhost:3000/api/v1
    description: Local development
security:
  - apiKeyAuth: []
  - bearerAuth: []
...
paths:
...
  /analytics/kpis:
    get:
      tags: [analytics]
      summary: Headline KPIs for a date range
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses: { '200': { description: KPI snapshot } }
  /analytics/series:
    get:
      tags: [analytics]
      summary: Revenue time-series + top products
      parameters:
        - name: bucket
          in: query
          schema: { type: string, enum: [day, month] }
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses: { '200': { description: Series } }
  /analytics/reports/{type}:
    get:
      tags: [analytics]
      summary: Build a tabular report by type
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Report table }
        '400': { $ref: '#/components/responses/ValidationError' }
  /analytics/reports/{type}/export:
    get:
      tags: [analytics]
      summary: Export a report as CSV or XLSX
      parameters:
        - name: type
          in: path
          required: true
          schema: { type: string }
        - name: format
          in: query
          schema: { type: string, enum: [csv, xlsx], default: csv }
      responses:
        '200':
          description: File download
          content:
            text/csv: {}
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet: {}
  /analytics/forecast:
    get:
      tags: [analytics]
      summary: Local demand forecast + reorder suggestions
      parameters:
        - name: lookback
          in: query
          schema: { type: integer, default: 90 }
        - name: horizon
          in: query
          schema: { type: integer, default: 30 }
      responses: { '200': { description: Forecast } }

  # ==================== AI ====================
...
  /admin/users:
    get:
      tags: [admin]
      summary: List tenant users
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/search'
      responses: { '200': { description: Users } }
    post:
      tags: [admin]
      summary: Create a user with role codes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, fullName, password]
              properties:
                email: { type: string, format: email }
                fullName: { type: string }
                password: { type: string, minLength: 6 }
                roleCodes:
                  type: array
                  items: { type: string }
      responses: { '201': { description: Created } }
  /admin/roles:
    get:
      tags: [admin]
      summary: List roles with permissions
      responses: { '200': { description: Roles } }
    post:
      tags: [admin]
      summary: Create a custom role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, name, permissions]
              properties:
                code: { type: string }
                name: { type: string }
                description: { type: string }
                permissions:
                  type: array
                  items: { type: string }
      responses: { '201': { description: Created } }
  /admin/invitations:
    get:
      tags: [admin]
      summary: List pending invitations
      responses: { '200': { description: Invitations } }
    post:
      tags: [admin]
      summary: Invite a user by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, roleCodes]
              properties:
                email: { type: string, format: email }
                roleCodes:
                  type: array
                  items: { type: string }
      responses: { '201': { description: Invitation created } }
  /admin/audit:
    get:
      tags: [admin]
      summary: Audit log (who / what / when)
      parameters:
        - $ref: '#/components/parameters/page'
      responses: { '200': { description: Audit entries } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Access token from POST /auth/login, sent as `Authorization: Bearer <token>`.'
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: 'Developer API key (ttr_live_…) created in the app under Управление → API-ключи. Recommended for server-to-server integrations. May also be sent as `Authorization: Bearer ttr_live_…`.'
  parameters:
...
    page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
...
    search:
      name: search
      in: query
      schema: { type: string }
  responses:
...
    ValidationError:
      description: Validation or business-rule error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            code: { type: string, example: VALIDATION_ERROR }
            message: { type: string }
            details: { type: object, nullable: true }
...